Add recursive uncompression utility

+++ src/main/java/org/nemac/util/Compression.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
 import org.apache.commons.io.FileUtils;
 
@@ -11,7 +12,7 @@ public class Compression {
 
     private static final String COMPRESSION_EXTENSION = ".gz";
 
-    public static String compress(String sourceFileName) throws IOException {
+    public static String compress(String sourceFileName, boolean clean) throws IOException {
         String outputFileName = sourceFileName + COMPRESSION_EXTENSION;
 
         File sourceFile = new File(sourceFileName);
@@ -28,8 +29,35 @@ public class Compression {
         in.close();
         out.close();
 
-        // cleanup source file
-        FileUtils.forceDelete(sourceFile);
+        if (clean) {
+            // cleanup source file
+            FileUtils.forceDelete(sourceFile);
+        }
+
+        return outputFileName;
+    }
+    
+    public static String uncompress(String sourceFileName, boolean cleanup) throws IOException {
+        int suffixDelimiterPosition = sourceFileName.lastIndexOf(COMPRESSION_EXTENSION);
+        String outputFileName = sourceFileName.substring(0, suffixDelimiterPosition);
+
+        File sourceFile = new File(sourceFileName);
+        GZIPInputStream inputZip = new GZIPInputStream(new FileInputStream(sourceFile));
+        FileOutputStream outputFile = new FileOutputStream(outputFileName);
+        
+        byte[] buffer = new byte[4096];
+        int len;
+        while ((len = inputZip.read(buffer)) > 0) {
+            outputFile.write(buffer, 0, len);
+        }
+
+        inputZip.close();
+        outputFile.close();
+
+        // cleanup source zip
+        if (cleanup) {
+            FileUtils.forceDelete(sourceFile);
+        }
 
         return outputFileName;
     }
